refactor and fix modal
