+++ components/Chat/Message/Message.tsx
@@ -29,7 +29,7 @@ import Animated, {
 } from "react-native-reanimated";
 
 import ChatMessageActions from "./MessageActions";
-import { ChatMessageReactions } from "./MessageReactions";
+import { ChatMessageReactions } from "./MessageReactions/MessageReactions";
 import MessageStatus from "./MessageStatus";
 import {
   currentAccount,
