+++ components/Chat/Message/MessageReactions/MessageReactions.types.ts
@@ -1,0 +1,5 @@
+import { MessageReaction } from "../../../../utils/reactions";
+
+export type MessageReactions = {
+  [senderAddress: string]: MessageReaction[];
+};
