+++ components/Chat/Message/MessageReactions/MessageReactionsDrawer/MessageReactions.store.tsx
@@ -1,0 +1,13 @@
+import { create } from "zustand";
+
+import { MessageToDisplay } from "../../Message";
+
+export interface IMessageReactionsStore {
+  message: MessageToDisplay | null;
+}
+
+export const useMessageReactionsStore = create<IMessageReactionsStore>(
+  (set, get) => ({
+    message: null,
+  })
+);
